Moved compat.py
